requirements.txt
langchain>=0.3.21
langchain-groq>=0.3.1
langchain-ollama>=0.3.0
pydantic>=2.11.1
python-dotenv>=1.1.0
rich>=13.9.4
streamlit>=1.44.0
watchdog>=6.0.0